test(goldenmatch-js): FS-NE cross-surface parity + homonym E2E
T6 closes the TS FS-NE port: Python-trained fixture re-scored in TS at
FULL float equality (EM-learned + penalty_bits variants), the #1764
homonym success bar as a TS E2E (traps merge without NE, separate with
it, dups keep merging), CHANGELOG/README capability notes, and the T5
review ride-alongs (probe citation, replicate() blocking note,
monotonicity-repair breadcrumb, NE M-step counts -> newM rename).
T5 trainEM suites split out to fs-ne-training.test.ts.
