+++ packages/typescript/goldenmatch/tests/parity/fsNegativeEvidence.parity.test.ts
@@ -1,0 +1,111 @@
+/**
+ * FS negative-evidence cross-surface SCORING parity (TS FS-NE port, T6).
+ *
+ * Fixture: tests/parity/fs-negative-evidence-fixtures.json, generated by a
+ * Python script (scratchpad, not committed — provenance in the fixture's
+ * `generator` field): `train_em` on the random-pair path (defaults, seed 42,
+ * GOLDENMATCH_FS_CALIBRATED / GOLDENMATCH_FS_MONOTONIC unset) + the scalar
+ * `score_pair_probabilistic`, for an EM-learned NE variant and a
+ * penalty_bits variant of the same 80-row homonym-shaped dataset.
+ *
+ * This pins SCORING parity: the Python-trained model (EMResult.to_dict())
+ * is re-scored in TS — config through `parseConfig` (exercising the T1/T2
+ * loader + validation-matrix path on real Python `model_dump` output),
+ * model through `emResultFromJson`, each pair through
+ * `scoreProbabilisticPair`. Cross-surface TRAINING comparison is not
+ * meaningful here (TS trainEM has no blocked-pairs path and the RNGs
+ * differ above the enumerate-all-pairs regime), so the fixture ships the
+ * trained model.
+ *
+ * Equality is FULL float precision (`toBe`, the PR #1755 probe standard):
+ * comparison vectors are integer levels, the weights are the exact doubles
+ * from the JSON round-trip, and both surfaces accumulate in matchkey-field
+ * order then NE order — so the arithmetic is bit-identical. A mismatch here
+ * is a semantic divergence to INVESTIGATE, never a tolerance to loosen.
+ *
+ * Pair mix: unfired dup (phones agree), fired homonym trap, null-phone
+ * inconclusive, empty-after-digits_only inconclusive, and cross pairs with
+ * disagreeing regular fields.
+ */
+import { describe, it, expect } from "vitest";
+import { readFileSync } from "node:fs";
+import { fileURLToPath } from "node:url";
+import { dirname, join } from "node:path";
+import {
+  emResultFromJson,
+  scoreProbabilisticPair,
+  validateEmResultFor,
+} from "../../src/core/probabilistic.js";
+import { parseConfig } from "../../src/core/config/loader.js";
+import type { Row } from "../../src/core/types.js";
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+interface PairScore {
+  a_idx: number;
+  b_idx: number;
+  score: number;
+}
+interface Variant {
+  config: Record<string, unknown>;
+  em_model: Record<string, unknown>;
+  expected_pair_scores: PairScore[];
+}
+interface Fixture {
+  generator: string;
+  rows: Record<string, unknown>[];
+  variants: Record<string, Variant>;
+}
+
+const fixture = JSON.parse(
+  readFileSync(join(__dirname, "fs-negative-evidence-fixtures.json"), "utf-8"),
+) as Fixture;
+
+const rows: Row[] = fixture.rows as Row[];
+
+describe("FS negative-evidence scoring parity vs Python", () => {
+  for (const [name, variant] of Object.entries(fixture.variants)) {
+    describe(`variant: ${name}`, () => {
+      // Python `GoldenMatchConfig.model_dump(exclude_none=True)` output fed
+      // straight into the loader — the T1/T2 parsing + NE validation matrix
+      // must accept the real Python wire shape.
+      const config = parseConfig(variant.config);
+      const mk = config.matchkeys?.[0];
+      const em = emResultFromJson(variant.em_model);
+
+      it("loader parses the probabilistic+NE matchkey (T1/T2)", () => {
+        expect(mk?.type).toBe("probabilistic");
+        const ne = mk?.negativeEvidence ?? [];
+        expect(ne).toHaveLength(1);
+        expect(ne[0]?.field).toBe("phone");
+        expect(ne[0]?.transforms).toEqual(["digits_only"]);
+        if (name === "penalty_bits") {
+          expect(ne[0]?.penaltyBits).toBe(6.0);
+          // penalty_bits NE skips EM: no __ne__ entry in the trained model.
+          expect(em.matchWeights["__ne__phone"]).toBeUndefined();
+        } else {
+          expect(ne[0]?.penaltyBits).toBeUndefined();
+          expect(em.matchWeights["__ne__phone"]).toHaveLength(2);
+          expect(em.matchWeights["__ne__phone"]![1]).toBe(0.0);
+        }
+      });
+
+      it("Python-trained model validates against the loaded matchkey", () => {
+        expect(() => validateEmResultFor(em, mk!)).not.toThrow();
+      });
+
+      it("scoreProbabilisticPair equals Python to FULL float precision", () => {
+        for (const p of variant.expected_pair_scores) {
+          const got = scoreProbabilisticPair(rows[p.a_idx]!, rows[p.b_idx]!, mk!, em);
+          // toBe, not toBeCloseTo: same doubles, same accumulation order.
+          expect(got, `pair (${p.a_idx}, ${p.b_idx})`).toBe(p.score);
+        }
+      });
+
+      it("the veto is visible: trap pair scores strictly below the dup pair", () => {
+        const [dup, trap] = variant.expected_pair_scores;
+        expect(trap!.score).toBeLessThan(dup!.score);
+      });
+    });
+  }
+});
